Refactoring. Add test. add useWidgetExportActionComponent hook
